+++ encryption/des_ecb.go
