+++ pkg/aes/aes_internal.go
@@ -4,8 +4,8 @@ import (
 	"crypto/aes"
 	"encoding/binary"
 	"errors"
+	"github.com/moov-io/dukpt/encryption"
 	"github.com/moov-io/dukpt/pkg"
-	"github.com/moov-io/dukpt/pkg/encryption"
 )
 
 const (
