+++ pkg/des/des_internal.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"crypto/des"
 	"fmt"
-	"github.com/moov-io/dukpt/pkg/encryption"
+	"github.com/moov-io/dukpt/encryption"
 	"math/big"
 )
